# make encoding.MarshalVarInt64s faster