Use `Switch/case` to replace `for`.
go compiler might be use jump table to replace compare.

Before:
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<6)-1
BenchmarkMarshalVarInt64s/up-to-(1<<6)-1
BenchmarkMarshalVarInt64s/up-to-(1<<6)-1-10              1551774               686.0 ns/op      11661.55 MB/s          0 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<13)-1
BenchmarkMarshalVarInt64s/up-to-(1<<13)-1
BenchmarkMarshalVarInt64s/up-to-(1<<13)-1-10              703304              1759 ns/op        4547.44 MB/s           0 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<27)-1
BenchmarkMarshalVarInt64s/up-to-(1<<27)-1
BenchmarkMarshalVarInt64s/up-to-(1<<27)-1-10              362886              3343 ns/op        2393.31 MB/s           4 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<63)-1
BenchmarkMarshalVarInt64s/up-to-(1<<63)-1
BenchmarkMarshalVarInt64s/up-to-(1<<63)-1-10              212334              6472 ns/op        1236.18 MB/s          18 B/op          0 allocs/op
PASS
ok      [github.com/VictoriaMetrics/VictoriaMetrics/lib/encoding](http://github.com/VictoriaMetrics/VictoriaMetrics/lib/encoding) 7.274s

After:
BenchmarkMarshalVarInt64s/up-to-(1<<6)-1
BenchmarkMarshalVarInt64s/up-to-(1<<6)-1-10               712990              1443 ns/op        5545.80 MB/s           0 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<13)-1
BenchmarkMarshalVarInt64s/up-to-(1<<13)-1
BenchmarkMarshalVarInt64s/up-to-(1<<13)-1-10              796993              1475 ns/op        5425.56 MB/s           0 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<27)-1
BenchmarkMarshalVarInt64s/up-to-(1<<27)-1
BenchmarkMarshalVarInt64s/up-to-(1<<27)-1-10              673441              1757 ns/op        4552.40 MB/s           2 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<63)-1
BenchmarkMarshalVarInt64s/up-to-(1<<63)-1
BenchmarkMarshalVarInt64s/up-to-(1<<63)-1-10              439021              2738 ns/op        2921.60 MB/s           8 B/op          0 allocs/op

====

BTW, c version is faster 40 times than golang version.
So go assemly version is on the way.
I will commit faster version next time.