@ahfuzhang 2024-01-31
> @ahfuzhang , thanks for the optimization! It would be great optimizing the case for small integers with up to `1<<6-1` values, since this case is very common in real world. According to the provided benchmark results, the performance for this case has been dropped from 11.7Gbytes/sec to 5.5Gbytes/sec.

because I don't add fast path for 1 byte marshal.
Wait a moment pls, I add it and test again.