@ahfuzhang 2024-01-31
New bechmark:

=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<6)-1
BenchmarkMarshalVarInt64s/up-to-(1<<6)-1
BenchmarkMarshalVarInt64s/up-to-(1<<6)-1-10              1795152               659.3 ns/op      12133.53 MB/s          0 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<13)-1
BenchmarkMarshalVarInt64s/up-to-(1<<13)-1
BenchmarkMarshalVarInt64s/up-to-(1<<13)-1-10              673744              1614 ns/op        4958.10 MB/s           0 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<27)-1
BenchmarkMarshalVarInt64s/up-to-(1<<27)-1
BenchmarkMarshalVarInt64s/up-to-(1<<27)-1-10              645064              1742 ns/op        4593.73 MB/s           2 B/op          0 allocs/op
=== RUN   BenchmarkMarshalVarInt64s/up-to-(1<<63)-1
BenchmarkMarshalVarInt64s/up-to-(1<<63)-1
BenchmarkMarshalVarInt64s/up-to-(1<<63)-1-10              435181              2515 ns/op        3180.43 MB/s           8 B/op          0 allocs/op
PASS
ok      [github.com/VictoriaMetrics/VictoriaMetrics/lib/encoding](http://github.com/VictoriaMetrics/VictoriaMetrics/lib/encoding) 5.385s

And new code is:
```go
func MarshalVarInt64s(dst []byte, vs []int64) []byte {
	for _, v := range vs {
		n := uint64((v << 1) ^ (v >> 63))
		if n < (1 << 7) {
			dst = append(dst, byte(n))
			continue
		}
		switch (64 - bits.LeadingZeros64(n>>1)) / 7 {
		case 0:
			dst = append(dst, byte(n))
		case 1:
			dst = append(dst, byte(n|0x80), byte(n>>7))
		case 2:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte(n>>14))
		case 3:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte((n>>14)|0x80), byte(n>>21))
		case 4:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte((n>>14)|0x80), byte((n>>21)|0x80), byte(n>>28))
		case 5:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte((n>>14)|0x80), byte((n>>21)|0x80), byte(n>>28|0x80), byte(n>>35))
		case 6:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte((n>>14)|0x80), byte((n>>21)|0x80), byte(n>>28|0x80), byte(n>>35|0x80), byte(n>>42))
		case 7:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte((n>>14)|0x80), byte((n>>21)|0x80), byte(n>>28|0x80), byte(n>>35|0x80), byte(n>>42|0x80), byte(n>>49))
		case 8:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte((n>>14)|0x80), byte((n>>21)|0x80), byte(n>>28|0x80), byte(n>>35|0x80), byte(n>>42|0x80), byte(n>>49|0x80), byte(n>>56))
		case 9:
			fallthrough
		default:
			dst = append(dst, byte(n|0x80), byte((n>>7)|0x80), byte((n>>14)|0x80), byte((n>>21)|0x80), byte(n>>28|0x80), byte(n>>35|0x80), byte(n>>42|0x80), byte(n>>49|0x80), byte(n>>56|0x80), byte(n>>63))
		}
	}
	return dst
}
```