@ahfuzhang 2024-01-31
And I make `UnmarshalVarInt64s` 28% faster.

Before:
=== RUN   BenchmarkUnmarshalVarInt64s/up-to-(1<<6)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<6)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<6)-1-10            1579142               793.7 ns/op      10079.69 MB/s          0 B/op       0 allocs/op
=== RUN   BenchmarkUnmarshalVarInt64s/up-to-(1<<13)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<13)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<13)-1-10            568491              1816 ns/op        4404.86 MB/s           0 B/op       0 allocs/op
=== RUN   BenchmarkUnmarshalVarInt64s/up-to-(1<<27)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<27)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<27)-1-10            231183              4840 ns/op        1652.94 MB/s           0 B/op       0 allocs/op
=== RUN   BenchmarkUnmarshalVarInt64s/up-to-(1<<63)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<63)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<63)-1-10            124256              8935 ns/op         895.37 MB/s           0 B/op       0 allocs/op

After:
BenchmarkUnmarshalVarInt64s/up-to-(1<<6)-1-10            1543092               796.8 ns/op      10039.85 MB/s          0 B/op          0 allocs/op
=== RUN   BenchmarkUnmarshalVarInt64s/up-to-(1<<13)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<13)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<13)-1-10           1000000              1123 ns/op        7125.49 MB/s           0 B/op          0 allocs/op
=== RUN   BenchmarkUnmarshalVarInt64s/up-to-(1<<27)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<27)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<27)-1-10            282601              3802 ns/op        2104.30 MB/s           0 B/op          0 allocs/op
=== RUN   BenchmarkUnmarshalVarInt64s/up-to-(1<<63)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<63)-1
BenchmarkUnmarshalVarInt64s/up-to-(1<<63)-1-10            179688              6460 ns/op        1238.30 MB/s           0 B/op          0 allocs/op

Code is:
```go
// UnmarshalVarInt64s unmarshals len(dst) int64 values from src to dst
// and returns the remaining tail from src.
func UnmarshalVarInt64s(dst []int64, src []byte) ([]byte, error) {
	idx := uint(0)
	for i := range dst {
		if idx >= uint(len(src)) {
			return nil, fmt.Errorf("cannot unmarshal varint from empty data")
		}
		c := src[idx]
		idx++
		if c < 0x80 {
			// Fast path
			v := int8(c>>1) ^ (int8(c<<7) >> 7) // zig-zag decoding without branching.
			dst[i] = int64(v)
			continue
		}
		if idx < uint(len(src)) && src[idx] < 0x80 {
			// Fast path, for 2 bytes
			n := uint64(c&0x7f) | (uint64(src[idx]&0x7f) << 7)
			dst[i] = int64(n>>1) ^ (int64(n<<63) >> 63)
			idx++
			continue
		}
		j := idx + 1
		for ; j < uint(len(src)); j++ { // find end loc
			if src[j] < 0x80 {
				break
			}
		}
		if j-idx > 10 {
			return nil, fmt.Errorf("cannot unmarshal varint, buffer too long, len=%d", j-idx)
		}
		dst[i] = unmarshalVarInt64ForOne(src[idx-1 : j+1])
		idx = j + 1
	}
	return src[idx:], nil
}

func unmarshalVarInt64ForOne(buf []byte) int64 {
	var n uint64
	switch len(buf) {
	case 1:
		n = uint64(buf[0] & 0x7F)
	case 2:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7)
	case 3:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14)
	case 4:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14) | (uint64(buf[3]&0x7F) << 21)
	case 5:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14) | (uint64(buf[3]&0x7F) << 21) | (uint64(buf[4]&0x7F) << 28)
	case 6:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14) | (uint64(buf[3]&0x7F) << 21) | (uint64(buf[4]&0x7F) << 28) | (uint64(buf[5]&0x7F) << 35)
	case 7:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14) | (uint64(buf[3]&0x7F) << 21) | (uint64(buf[4]&0x7F) << 28) | (uint64(buf[5]&0x7F) << 35) | (uint64(buf[6]&0x7F) << 42)
	case 8:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14) | (uint64(buf[3]&0x7F) << 21) | (uint64(buf[4]&0x7F) << 28) | (uint64(buf[5]&0x7F) << 35) | (uint64(buf[6]&0x7F) << 42) | (uint64(buf[7]&0x7F) << 49)
	case 9:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14) | (uint64(buf[3]&0x7F) << 21) | (uint64(buf[4]&0x7F) << 28) | (uint64(buf[5]&0x7F) << 35) | (uint64(buf[6]&0x7F) << 42) | (uint64(buf[7]&0x7F) << 49) | (uint64(buf[8]&0x7F) << 56)
	case 10:
		n = uint64(buf[0]&0x7F) | (uint64(buf[1]&0x7F) << 7) | (uint64(buf[2]&0x7F) << 14) | (uint64(buf[3]&0x7F) << 21) | (uint64(buf[4]&0x7F) << 28) | (uint64(buf[5]&0x7F) << 35) | (uint64(buf[6]&0x7F) << 42) | (uint64(buf[7]&0x7F) << 49) | (uint64(buf[8]&0x7F) << 56) | (uint64(buf[9]&0x7F) << 63)
	default:
		panic("impossible error: buf length must be 1 to 10")
	}
	return int64(n>>1) ^ (int64(n<<63) >> 63)
}

```